pyproject.toml
[project]
name = "roof-ranker"
version = "0.1.0"
description = "Solar potential analysis for institutional buildings using satellite imagery"
readme = "README.md"
authors = [{ name = "dudosya", email = "kenaykay@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "geopy>=2.4.0",
    "googlemaps>=4.10.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "segment-anything>=1.0.0",
    "plotly>=6.5.0",
    "streamlit-drawable-canvas>=0.9.3",
    "streamlit-image-coordinates>=0.4.0",
]

[project.scripts]
roof-ranker = "roof_ranker.app:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
]
